requirements.txt
streamlit
pandas
langchain
langchain_experimental
python-dotenv
numpy
openai
python-dotenv
seaborn